# lib/schema/gen/go/api/core/v1/event.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.7
// 	protoc        (unknown)
// source: api/core/v1/event.proto

package corev1

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type EventName int32

const (
	EventName_EVENT_NAME_UNSPECIFIED EventName = 0
	EventName_EVENT_NAME_KEEP_ALIVE  EventName = 1
	EventName_EVENT_NAME_TICK        EventName = 2
)

// Enum value maps for EventName.
var (
	EventName_name = map[int32]string{
		0: "EVENT_NAME_UNSPECIFIED",
		1: "EVENT_NAME_KEEP_ALIVE",
		2: "EVENT_NAME_TICK",
	}
	EventName_value = map[string]int32{
		"EVENT_NAME_UNSPECIFIED": 0,
		"EVENT_NAME_KEEP_ALIVE":  1,
		"EVENT_NAME_TICK":        2,
	}
)

func (x EventName) Enum() *EventName {
	p := new(EventName)
	*p = x
	return p
}

func (x EventName) String() string {
	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
}

func (EventName) Descriptor() protoreflect.EnumDescriptor {
	return file_api_core_v1_event_proto_enumTypes[0].Descriptor()
}

func (EventName) Type() protoreflect.EnumType {
	return &file_api_core_v1_event_proto_enumTypes[0]
}

func (x EventName) Number() protoreflect.EnumNumber {
	return protoreflect.EnumNumber(x)
}

// Deprecated: Use EventName.Descriptor instead.
func (EventName) EnumDescriptor() ([]byte, []int) {
	return file_api_core_v1_event_proto_rawDescGZIP(), []int{0}
}

type Event struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Name          EventName              `protobuf:"varint,1,opt,name=name,proto3,enum=api.core.v1.EventName" json:"name,omitempty"`
	Tick          *Tick                  `protobuf:"bytes,2,opt,name=tick,proto3" json:"tick,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *Event) Reset() {
	*x = Event{}
	mi := &file_api_core_v1_event_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *Event) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*Event) ProtoMessage() {}

func (x *Event) ProtoReflect() protoreflect.Message {
	mi := &file_api_core_v1_event_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use Event.ProtoReflect.Descriptor instead.
func (*Event) Descriptor() ([]byte, []int) {
	return file_api_core_v1_event_proto_rawDescGZIP(), []int{0}
}

func (x *Event) GetName() EventName {
	if x != nil {
		return x.Name
	}
	return EventName_EVENT_NAME_UNSPECIFIED
}

func (x *Event) GetTick() *Tick {
	if x != nil {
		return x.Tick
	}
	return nil
}

type Tick struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Provider      string                 `protobuf:"bytes,1,opt,name=provider,proto3" json:"provider,omitempty"`
	Ticker        string                 `protobuf:"bytes,2,opt,name=ticker,proto3" json:"ticker,omitempty"`
	Price         float64                `protobuf:"fixed64,3,opt,name=price,proto3" json:"price,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *Tick) Reset() {
	*x = Tick{}
	mi := &file_api_core_v1_event_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *Tick) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*Tick) ProtoMessage() {}

func (x *Tick) ProtoReflect() protoreflect.Message {
	mi := &file_api_core_v1_event_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use Tick.ProtoReflect.Descriptor instead.
func (*Tick) Descriptor() ([]byte, []int) {
	return file_api_core_v1_event_proto_rawDescGZIP(), []int{1}
}

func (x *Tick) GetProvider() string {
	if x != nil {
		return x.Provider
	}
	return ""
}

func (x *Tick) GetTicker() string {
	if x != nil {
		return x.Ticker
	}
	return ""
}

func (x *Tick) GetPrice() float64 {
	if x != nil {
		return x.Price
	}
	return 0
}

var File_api_core_v1_event_proto protoreflect.FileDescriptor

const file_api_core_v1_event_proto_rawDesc = "" +
	"\n" +
	"\x17api/core/v1/event.proto\x12\vapi.core.v1\"Z\n" +
	"\x05Event\x12*\n" +
	"\x04name\x18\x01 \x01(\x0e2\x16.api.core.v1.EventNameR\x04name\x12%\n" +
	"\x04tick\x18\x02 \x01(\v2\x11.api.core.v1.TickR\x04tick\"P\n" +
	"\x04Tick\x12\x1a\n" +
	"\bprovider\x18\x01 \x01(\tR\bprovider\x12\x16\n" +
	"\x06ticker\x18\x02 \x01(\tR\x06ticker\x12\x14\n" +
	"\x05price\x18\x03 \x01(\x01R\x05price*W\n" +
	"\tEventName\x12\x1a\n" +
	"\x16EVENT_NAME_UNSPECIFIED\x10\x00\x12\x19\n" +
	"\x15EVENT_NAME_KEEP_ALIVE\x10\x01\x12\x13\n" +
	"\x0fEVENT_NAME_TICK\x10\x02B5Z3market-tracker/lib/schema/gen/go/api/core/v1;corev1b\x06proto3"

var (
	file_api_core_v1_event_proto_rawDescOnce sync.Once
	file_api_core_v1_event_proto_rawDescData []byte
)

func file_api_core_v1_event_proto_rawDescGZIP() []byte {
	file_api_core_v1_event_proto_rawDescOnce.Do(func() {
		file_api_core_v1_event_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_api_core_v1_event_proto_rawDesc), len(file_api_core_v1_event_proto_rawDesc)))
	})
	return file_api_core_v1_event_proto_rawDescData
}

var file_api_core_v1_event_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
var file_api_core_v1_event_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_api_core_v1_event_proto_goTypes = []any{
	(EventName)(0), // 0: api.core.v1.EventName
	(*Event)(nil),  // 1: api.core.v1.Event
	(*Tick)(nil),   // 2: api.core.v1.Tick
}
var file_api_core_v1_event_proto_depIdxs = []int32{
	0, // 0: api.core.v1.Event.name:type_name -> api.core.v1.EventName
	2, // 1: api.core.v1.Event.tick:type_name -> api.core.v1.Tick
	2, // [2:2] is the sub-list for method output_type
	2, // [2:2] is the sub-list for method input_type
	2, // [2:2] is the sub-list for extension type_name
	2, // [2:2] is the sub-list for extension extendee
	0, // [0:2] is the sub-list for field type_name
}

func init() { file_api_core_v1_event_proto_init() }
func file_api_core_v1_event_proto_init() {
	if File_api_core_v1_event_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_api_core_v1_event_proto_rawDesc), len(file_api_core_v1_event_proto_rawDesc)),
			NumEnums:      1,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_api_core_v1_event_proto_goTypes,
		DependencyIndexes: file_api_core_v1_event_proto_depIdxs,
		EnumInfos:         file_api_core_v1_event_proto_enumTypes,
		MessageInfos:      file_api_core_v1_event_proto_msgTypes,
	}.Build()
	File_api_core_v1_event_proto = out.File
	file_api_core_v1_event_proto_goTypes = nil
	file_api_core_v1_event_proto_depIdxs = nil
}